strategy circle
As we have kicked off work with the advisors, we need to reform this role. I'd like to make a circle out of this role, with roles both for the external advisors and some of our professional partners. I think advisors should work with us in the way we work internally.